IDAM-107 - refactoring namespaces and models
