+++ NICE.Identity.Authorisation.WebAPI/DataModels/Websites.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using NICE.Identity.Authorisation.WebAPI.DataModels;
 
-namespace NICE.Identity.Authorisation.WebAPI.Models
+namespace NICE.Identity.Authorisation.WebAPI.DataModels
 {
     public partial class Websites
     {
