--- NICE.Identity.Authorisation.WebAPI/Domain/Role.cs
@@ -1,11 +1,0 @@
-﻿namespace NICE.Identity.Authorisation.WebAPI.Domain
-{
-    public class Role
-    {
-        /*TODO Replace with Entity Framework domain model*/
-
-        public string Id { get; set; }
-
-        public string Name { get; set; }
-    }
-}
