+++ NICE.Identity.Test/IntegrationTests/ClaimsTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using NICE.Identity.Authorisation.WebAPI.Models;
 using NICE.Identity.Test.Infrastructure;
 using Xunit;
